Move nats controllers into a new file
